md2word-api/requirements.txt
# MD2Word API - WeldTech Solutions
# Dependencias de Python
# ======================================

# Core dependencies
python-docx==1.1.0
pypandoc==1.11
pyyaml==6.0.1

# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# CLI dependencies
click==8.1.7
rich==13.7.0

# Development dependencies (opcional)
# pytest==7.4.3
# black==23.12.0
# flake8==6.1.0
# mypy==1.7.1

